+++ public/assets/application-a5804efe17369f430214a254e3b8b991.js
@@ -13431,14 +13431,16 @@ for(d=[],m=a=b=this.ymin,c=this.ymax;k>0?c>=a:a>=c;m=a+=k)d.push(m);return d}.ca
             var panel = $(this).closest(".panel"),
                 postId = panel.data("post-id");
 
-            $.ajax({
-                url: "/posts/" + postId,
-                method: "DELETE",
+            if (confirm("Are you sure you want to delete this post?")) {
+                $.ajax({
+                    url: "/posts/" + postId,
+                    method: "DELETE",
 
-                success: function () {
-                    panel.remove();
-                }
-            });
+                    success: function () {
+                        panel.remove();
+                    }
+                });
+            }
         });
     });
 }());
